No longer using the -I in traceroute & added nice
The “-I” option makes the traceroute run a little faster, but it takes
super-user privileges to do so. This means it can’t be run on shared
hosting, so I removed it.
The addition of a the “nice” command makes sure that we are using
running at the lowest process which will help run on a busy system.